add mocks fonder & move offer mocks
