+++ src/components/app/app.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import renderer from "react-test-renderer";
 import App from "./app";
-import {OFFERS} from "../../mocks";
+import {OFFERS} from "../../mocks/offers";
 
 it(`App component render correct`, () => {
   const app = renderer
