+++ src/index.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './components/app/app';
-import {OFFERS} from "./mocks";
+import {OFFERS} from "./mocks/offers";
 
 const rootElement = document.querySelector(`#root`);
